third_party/WebKit/LayoutTests/http/tests/origin_trials/webexposed/animationworklet-origin-trial-interfaces.html

<!DOCTYPE html>
<!-- Generate token with the command:
generate_token.py http://127.0.0.1:8000 AnimationWorklet --expire-timestamp=2000000000
-->
<meta http-equiv="origin-trial" content="AnRfiFGg50a1+vvDBVfbBNiiRFBBQDy+3oMlJw2pUNRb8tPyN2XX0kAJxiA0itCLJTM0DrsIfVYuv4ni0Lu6bQgAAABYeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiQW5pbWF0aW9uV29ya2xldCIsICJleHBpcnkiOiAyMDAwMDAwMDAwfQ==" />
<title>AnimationWorklet - interfaces exposed by origin trial</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/origin-trials-helper.js"></script>
<script>
test(t => {
  OriginTrialsHelper.check_interfaces(this,
      ['WorkletAnimation', 'ScrollTimeline']);
}, 'AnimationWorklet related interfaces in Origin-Trial enabled document.');

test(t => {
  assert_own_property(CSS, 'animationWorklet');
  OriginTrialsHelper.check_properties(this,
      {'WorkletAnimation': ['play', 'cancel', 'playState', 'timeline'],
       'ScrollTimeline': ['scrollSource', 'orientation', 'timeRange'],
      });
}, 'AnimationWorklet related properties in Origin-Trial enabled document.');

</script>
